T-9810773-Create-an-adaptive-layout-of-playing-field:refactoring the folder hierarchy in a project
