+++ package.json
@@ -14,6 +14,10 @@
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.5.2",
     "@types/node": "^16.18.23",
+    "@types/react": "18.2.0",
+    "@types/styled-components": "^5.1.26",
+    "babel-loader": "^8.2.2",
+    "eslint-plugin-material-ui": "^1.0.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-scripts": "5.0.1",
