--- src/components/cardsPanel/cardsPanel.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import { CardsPanelWrapper } from './cardsPanelStyles';
-
-export function CardsPanel() {
-  return (
-    <CardsPanelWrapper />
-  );
-}
