+++ src/components/mainPage/commonGameArea/cardsList/cardsList.tsx
