+++ src/components/mainPage/commonGameArea/cardsList/cardsListStyles.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { colors } from '../../theme/variables';
+import { colors } from '../../../../theme/variables';
 
 export const CardList = styled.div`
     grid-row: 2 / 3;
@@ -8,7 +8,7 @@ export const CardList = styled.div`
     justify-content: center;
     align-items: center;
     height: 20vh;
-    background-color: ${colors.CARDS_LIST_BACKGROUND_COLOR};
+    background-color: ${colors.cardsListBackgroundColor};
     overflow: auto;
 `;
 
@@ -16,7 +16,7 @@ export const Card = styled.div`
     width: 10vh;
     height: 15vh;
     margin: 0 0.2%;
-    background-color: ${colors.PRIMARY_COLOR};
+    background-color: ${colors.primaryColor};
 
   &:hover {
     transform: scale(1.2);
