+++ src/components/mainPage/commonGameArea/cardsList/index.ts
