+++ src/components/mainPage/commonGameArea/cardsPanel/cardsPanel.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { CardsPanelArea } from './cardsPanelStyles';
+
+export function CardsPanel() {
+  return (
+    <CardsPanelArea />
+  );
+}
