+++ src/components/mainPage/commonGameArea/cardsPanel/cardsPanelStyles.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
-import cardsPanelImage from '../assets/battlefield/cardsPanel-test.jpeg';
+import cardsPanelImage from '../../../../assets/battlefield/cardsPanel-test.jpeg';
 
-export const CardsPanelWrapper = styled.div`
+export const CardsPanelArea = styled.div`
   grid-row: 2 / 3;
   grid-column: 2 / 3;
   background-image: url(${cardsPanelImage});
