+++ src/components/mainPage/commonGameArea/cardsPanel/index.ts
