+++ src/components/mainPage/commonGameArea/commonGameAreaStyles.tsx
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+import { colors } from '../../../theme/variables';
+
+export const CommonGameAreaWrapper = styled.div`
+  display: grid;
+  grid-template-rows: 4fr 1fr;
+  grid-template-columns: 3fr 2fr;
+  grid-gap: 0.1vh;
+  height: 100vh;
+  width: 80%;
+  // border: 0.001em solid ${colors.secondaryColor};
+  // border-left: 0.001em solid ${colors.secondaryColor};
+  background-color: ${colors.commonGameAreaBackgroundColor};
+  box-sizing: border-box;
+  float: left;
+`;
