+++ src/components/mainPage/commonGameArea/gameArea/gameAreaSides/gameAreaSidesStyles/index.ts
@@ -1,0 +1,1 @@
+export * from './gameAreaSidesStyles';
