+++ src/components/mainPage/commonGameArea/gameArea/gameAreaSides/index.ts
@@ -1,0 +1,1 @@
+export * from './gameAreaSides';
