+++ src/components/mainPage/commonGameArea/gameArea/gameAreaWrapper/index.ts
@@ -1,0 +1,1 @@
+export * from './gameAreaWrapper';
