+++ src/components/mainPage/commonGameArea/gameArea/index.ts
