+++ src/components/mainPage/commonGameArea/gameCardArea/gameCardArea.tsx
