+++ src/components/mainPage/commonGameArea/gameCardArea/index.ts
