+++ src/components/mainPage/commonGameArea/index.ts
