+++ src/components/mainPage/mainPage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { OpponentsPanel } from '../opponentsPanel';
-import { CommonGameArea } from '../commonGameArea';
+import { OpponentsPanel } from './opponentsPanel';
+import { CommonGameArea } from './commonGameArea';
 
 export function MainPage() {
   return (
