+++ src/components/mainPage/opponentsPanel/index.ts
