+++ src/components/mainPage/opponentsPanel/opponentsPanel.tsx
