+++ src/theme/variables.tsx
@@ -1,12 +1,12 @@
 export const colors: Record<string, string> = {
-  TEXT_MAIN: '#EADDFF',
-  BACKGROUND: '#6868EA',
-  PRIMARY_COLOR: '#8B4513',
-  SECONDARY_COLOR: '#000000',
-  COMMON_GAME_AREA_BACKGROUND_COLOR: '#592d06',
-  CARDS_SECONDARY_BACKGROUND_COLOR: '#A0522D',
-  GAME_AREA_CIRCLE_BLOCK_HOVER_COLOR: '#FFFF00',
-  GAME_AREA_CIRCLE_BLOCK_BACKGROUND_COLOR: '#8D5120',
-  CARDS_LIST_BACKGROUND_COLOR: '#481601',
-  GAME_AREA_BACKGROUND_COLOR: '#503017',
+  textMain: '#EADDFF',
+  background: '#6868EA',
+  primaryColor: '#8B4513',
+  secondaryColor: '#000000',
+  commonGameAreaBackgroundColor: '#592D06',
+  cardsSecondaryBackgroundColor: '#A0522D',
+  gameAreaCircleBlockHoverColor: '#FFFF00',
+  gameAreaCircleBlockBackgroundColor: '#8D5120',
+  cardsListBackgroundColor: '#481601',
+  gameAreaBackgroundColor: '#503017',
 };
